Code is reverted back to JS rather than TS to stay within scope of the project
