+++ src/components/demoCharts.jsx
@@ -4,29 +4,7 @@ import { PieChart, Pie, Cell, ResponsiveContainer, BarChart, Bar, XAxis, Tooltip
 const COLORS = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042'];
 const RADIAN = Math.PI / 180;
 
-interface PieDataItem {
-  name: string;
-  value: number;
-  genre: string;
-}
-
-interface BarDataItem {
-  name: string;
-  achievements: number;
-}
-
-interface CustomizedLabelProps {
-  cx: number;
-  cy: number;
-  midAngle: number;
-  innerRadius: number;
-  outerRadius: number;
-  percent: number;
-  index: number;
-  payload: PieDataItem;
-}
-
-const renderCustomizedLabel = ({ cx, cy, midAngle, innerRadius, outerRadius, percent, payload }: CustomizedLabelProps) => {
+const renderCustomizedLabel = ({ cx, cy, midAngle, innerRadius, outerRadius, percent, payload }) => {
     const radius = innerRadius + (outerRadius - innerRadius) * 0.5;
     const x = cx + radius * Math.cos(-midAngle * RADIAN);
     const y = cy + radius * Math.sin(-midAngle * RADIAN);
@@ -38,8 +16,7 @@ const renderCustomizedLabel = ({ cx, cy, midAngle, innerRadius, outerRadius, per
     );
 };
 
-// Function to render a Pie Chart
-const renderDemoPieChart = (data: PieDataItem[]) => (
+const renderDemoPieChart = (data) => (
     <ResponsiveContainer width="100%" height="100%">
         <PieChart width={400} height={400}>
             <Pie
@@ -52,17 +29,19 @@ const renderDemoPieChart = (data: PieDataItem[]) => (
                 fill="#8884d8"
                 dataKey="value"
             >
-                {data.map((_, index) => (
+                {data.map((entry, index) => (
                     <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
                 ))}
             </Pie>
+            <Tooltip />
+            <Legend />
         </PieChart>
     </ResponsiveContainer>
 );
 
-const renderDemoBarChart = (data: BarDataItem[]) => (
-    <ResponsiveContainer width="100%" height={300}>
-        <BarChart data={data}>
+const renderDemoBarChart = (data) => (
+    <ResponsiveContainer width="100%" height="100%">
+        <BarChart width={500} height={300} data={data} margin={{ top: 5, right: 30, left: 20, bottom: 5 }}>
             <XAxis dataKey="name" />
             <Tooltip />
             <Legend />
@@ -71,9 +50,9 @@ const renderDemoBarChart = (data: BarDataItem[]) => (
     </ResponsiveContainer>
 );
 
-const renderSecondBarChart = (data: BarDataItem[]) => (
-    <ResponsiveContainer width="100%" height={300}>
-        <BarChart data={data}>
+const renderSecondBarChart = (data) => (
+    <ResponsiveContainer width="100%" height="100%">
+        <BarChart width={500} height={300} data={data} margin={{ top: 5, right: 30, left: 20, bottom: 5 }}>
             <XAxis dataKey="name" />
             <Tooltip />
             <Legend />
@@ -82,22 +61,25 @@ const renderSecondBarChart = (data: BarDataItem[]) => (
     </ResponsiveContainer>
 );
 
-interface DemoChartsState {
-    pieChart: React.ReactNode;
-    barChart: React.ReactNode;
-    secondBarChart: React.ReactNode;
-}
+class DemoCharts extends PureComponent {
+    constructor(props) {
+        super(props);
+        this.state = {
+            pieChart: null,
+            barChart: null,
+            secondBarChart: null
+        };
+    }
 
-class DemoCharts extends PureComponent<{}, DemoChartsState> {
-    render() {
-        const pieData: PieDataItem[] = [
-            { name: 'Group A', value: 400, genre: 'Shooter' },
-            { name: 'Group B', value: 300, genre: 'Strategy' },
-            { name: 'Group C', value: 300, genre: 'RPG' },
-            { name: 'Group D', value: 700, genre: 'Other' },
+    componentDidMount() {
+        const pieData = [
+            { name: 'Action', value: 40, genre: 'Action' },
+            { name: 'RPG', value: 30, genre: 'RPG' },
+            { name: 'Strategy', value: 20, genre: 'Strategy' },
+            { name: 'Others', value: 10, genre: 'Others' },
         ];
 
-        const barData: BarDataItem[] = [
+        const barData = [
             { name: '0:00', achievements: 53 },
             { name: '1:00', achievements: 30 },
             { name: '2:00', achievements: 55 },
@@ -124,7 +106,7 @@ class DemoCharts extends PureComponent<{}, DemoChartsState> {
             { name: '23:00', achievements: 44 },
         ];
 
-        const secondBarData: BarDataItem[] = [
+        const secondBarData = [
             { name: 'Sunday', achievements: 126 },
             { name: 'Monday', achievements: 72 },
             { name: 'Tuesday', achievements: 35 },
@@ -134,11 +116,30 @@ class DemoCharts extends PureComponent<{}, DemoChartsState> {
             { name: 'Saturday', achievements: 185 },
         ];
 
-        return {
+        this.setState({
             pieChart: renderDemoPieChart(pieData),
             barChart: renderDemoBarChart(barData),
-            secondBarChart: renderSecondBarChart(secondBarData),
-        };
+            secondBarChart: renderSecondBarChart(secondBarData)
+        });
+    }
+
+    render() {
+        const { pieChart, barChart, secondBarChart } = this.state;
+        return (
+            <div>
+                <div className="chart-container h-[250px] w-[100%]">
+                    {pieChart}
+                    <h2 className="text-xl text-center pt-2 pb-2 bg-base-100 mr-5 ml-5 rounded-xl mb-[10%]">You mainly play <span className='text-success font-bold'>Shooter</span> games</h2>
+                </div>
+
+                <div className="chart-container h-[800px] w-[100%] mt-[10%]">
+                    {barChart}
+                    <h2 className="text-xl text-center pt-2 pb-2 bg-base-100 mr-5 ml-5 rounded-xl mb-[10%]">You prefer to play games at <span className='text-success font-bold'>night</span></h2>
+                    {secondBarChart}
+                    <h2 className="text-xl text-center pt-2 pb-2 bg-base-100 mr-5 ml-5 rounded-xl mb-[10%]">You prefer to play games on <span className='text-success font-bold'>Saturday</span></h2>
+                </div>
+            </div>
+        );
     }
 }
